+++ src/GameMap/GameMap.java
@@ -856,38 +856,5 @@ public class GameMap {
         return mapObjects;
     }
 
-    public static void main(String[] args) {
-        try {
-            String address = "maps/testsForDifficultIncidents/0_test.map";
-
-            GameMap inputMap = GameMap.cutMapBetweenStartAndEnd(address, "is", "ie");
-            inputMap.setGrowth(GameMap.cutParamAfterWord(address, "Growth "));
-            inputMap.setRazors(GameMap.cutParamAfterWord(address, "Razors "));
-            inputMap.setFlooding(GameMap.cutParamAfterWord(address, "Flooding "));
-
-
-            //NextStep nextSteps[] = GameMap.cutSteps(address);
-            //for (NextStep nextStep : nextSteps)
-            //    inputMap.moveAllObjects(nextStep);
-            BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
-            String str;
-            System.out.println(inputMap.toString());
-            while (true) {
-                System.out.println("EnterSteps\n");
-                str = bufferedReader.readLine();
-                if (str.equals("break"))
-                    break;
-                NextStep nextSteps[] = GameMap.cutStepsFromString(str);
-                for (NextStep nextstep : nextSteps)
-                    inputMap.moveAllObjects(nextstep);
-                System.out.println(inputMap.toString() + "\n");
-                System.out.println("Score = " + inputMap.getScore());
-                System.out.println("Lambdas " + inputMap.getLamdasNumber() + "/" + inputMap.getMaxLambdasNumber());
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-    }
 
 }
